Support larger buffers and messages

See merge request espressif/esp-mqtt!56

+++ lib/include/mqtt_msg.h
@@ -1,5 +1,8 @@
 #ifndef MQTT_MSG_H
 #define MQTT_MSG_H
+#include <stdint.h>
+#include <stdbool.h>
+
 #include "mqtt_config.h"
 #ifdef  __cplusplus
 extern "C" {
@@ -69,7 +72,7 @@ typedef struct mqtt_connection {
 
     uint16_t message_id;
     uint8_t *buffer;
-    uint16_t buffer_length;
+    uint32_t buffer_length;
 
 } mqtt_connection_t;
 
@@ -117,13 +120,13 @@ static inline int mqtt_get_retain(uint8_t *buffer)
     return (buffer[0] & 0x01);
 }
 
-void mqtt_msg_init(mqtt_connection_t *connection, uint8_t *buffer, uint16_t buffer_length);
-bool mqtt_header_complete(uint8_t *buffer, uint16_t buffer_length);
-uint32_t mqtt_get_total_length(uint8_t *buffer, uint16_t length, int *fixed_size_len);
+void mqtt_msg_init(mqtt_connection_t *connection, uint8_t *buffer, uint32_t buffer_length);
+bool mqtt_header_complete(uint8_t *buffer, uint32_t buffer_length);
+uint32_t mqtt_get_total_length(uint8_t *buffer, uint32_t length, int *fixed_size_len);
 char *mqtt_get_publish_topic(uint8_t *buffer, uint32_t *length);
 char *mqtt_get_publish_data(uint8_t *buffer, uint32_t *length);
-uint16_t mqtt_get_id(uint8_t *buffer, uint16_t length);
-int mqtt_has_valid_msg_hdr(uint8_t *buffer, uint16_t length);
+uint16_t mqtt_get_id(uint8_t *buffer, uint32_t length);
+int mqtt_has_valid_msg_hdr(uint8_t *buffer, uint32_t length);
 
 mqtt_message_t *mqtt_msg_connect(mqtt_connection_t *connection, mqtt_connect_info_t *info);
 mqtt_message_t *mqtt_msg_publish(mqtt_connection_t *connection, const char *topic, const char *data, int data_length, int qos, int retain, uint16_t *message_id);
